+++ tests/testthat/test-004-pro_pipeline_single.R
@@ -56,8 +56,7 @@ test_that("pro_request_jsonl_parquet single identifier", {
   output_parquet <- output_jsonl |>
     pro_request_jsonl_parquet(
       output = output_parquet,
-      verbose = FALSE,
-      progress = TRUE
+      verbose = FALSE
     )
 
   # Check that the output file exists