+++ tests/testthat/test-005-pro_pipeline_search.R
@@ -57,8 +57,7 @@ test_that("pro_request_jsonl_parquet search `biodiversity AND finance`", {
   output_parquet <- output_jsonl |>
     pro_request_jsonl_parquet(
       output = output_parquet,
-      verbose = FALSE,
-      progress = TRUE
+      verbose = FALSE
     )
 
   # Check that the output file exists