+++ tests/testthat/test-008-pro_pipeline_parallel.R
@@ -96,8 +96,7 @@ test_that("pro_request_jsonl_parquet with subfolders", {
   output_parquet <- output_jsonl |>
     pro_request_jsonl_parquet(
       output = output_parquet,
-      verbose = FALSE,
-      progress = TRUE
+      verbose = FALSE
     )
 
   # Check that the output file exists